nxp: feature gate lpc55 gpio and pint to lpc55
